Move skill scanning helpers into common skillscan
