+++ cli/common/skillscan/cli_source.go
@@ -1,4 +1,4 @@
-package skills
+package skillscan
 
 import "path/filepath"
