+++ cli/common/skillscan/packages_test.go
@@ -1,4 +1,4 @@
-package clicore
+package skillscan
 
 import (
 	"os"
