+++ cli/project-runner/shared-inputs-stamp.json
@@ -1,4 +1,4 @@
 {
   "schemaVersion": 1,
-  "sharedInputsHash": "4d69c805f009f427144ecab4a78119d537e99c97"
+  "sharedInputsHash": "4927bc05b4c84ecfb1b01ff3d80974f20b12d8d0"
 }
