+++ scripts/stamp-release-inputs.sh
@@ -21,7 +21,7 @@ list_shared_common_inputs() {
     'cli/common/clicore/' \
     'cli/common/errors/' \
     'cli/common/project/' \
-    'cli/common/skills/' \
+    'cli/common/skillscan/' \
     'cli/common/tools/' \
     'cli/common/unityipc/' \
     'cli/common/unityprocess/' |
